src/gpsneo6/requirements.txt
pyserial
gps3
pynmea2
paho-mqtt